requirements.txt
PySide6>=6.7.0
ollama>=0.6.0
python-dotenv>=1.0.0
markdown>=3.5.0
sentence-transformers>=3.0.0
rank-bm25>=0.2.2
pypdf>=4.0.0
python-docx>=1.0.0
sentence-splitter>=1.4.0
fastapi>=0.115.0
uvicorn[standard]>=0.29.0